ENE-45 install pip emt CLI entrypoint
