+++ src/lib.rs
@@ -1,3 +1,4 @@
+pub mod cli;
 pub mod collectors;
 pub mod config;
 pub mod energy_group;
